server/requirements.txt
Django==5.2.1
# rest framework
djangorestframework==3.16.0
markdown==3.8  # Markdown support for the browsable API
django-filter==25.1  # Filtering support
gunicorn==23.0.0  # Production server
psycopg2-binary==2.9.10  # database
python-dotenv==1.1.0  # Environment variables
django-cors-headers==4.7.0  # CORS support
djangorestframework-simplejwt==5.5.0  # JWT authentication
django-countries==7.6.1  # Countries support